+++ include/SIMPLESDL/main.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+//Main header file for SIMPLESDL
+
+#include <SIMPLESDL/camera.hpp>
+#include <SIMPLESDL/collisions.hpp>
+#include <SIMPLESDL/engine.hpp>
+#include <SIMPLESDL/object.hpp>
