+++ include/SIMPLESDL/object.hpp
